+++ colormaps/src/main/java/colormaps/impl/CubeDiagonalCuttingRedCyan.java
@@ -26,10 +26,10 @@ import colormaps.ColorSpace;
 public class CubeDiagonalCuttingRedCyan extends AbstractColormap2D {
 
 	@Override
-	public Color getColor(float x, float y) {
+	public Color getColor(double x, double y) {
 		checkRanges(x, y);
 
-		return new Color(1 - y, x, y);
+		return new Color((float)(1 - y), (float)x, (float)y);
 	}
 
 	@Override
