+++ colormaps/src/main/java/colormaps/impl/RobertsonAndOCallaghan1.java
@@ -28,21 +28,21 @@ public class RobertsonAndOCallaghan1 extends AbstractColormap2D {
 	private FourCornersAnchorColorMapParameterizable[][] colorMaps;
 
 	@Override
-	public Color getColor(float x, float y) {
+	public Color getColor(double x, double y) {
 
 		if (colorMaps == null)
 			initializeColorMap();
 
 		checkRanges(x, y);
 
-		float fx = x;
+		double fx = x;
 		int indexX = 0;
 		while (fx > 1 / 3.0f) {
 			indexX++;
 			fx -= 1 / 3.0f;
 		}
 
-		float fy = y;
+		double fy = y;
 		int indexY = 0;
 		while (fy > 1 / 3.0f) {
 			indexY++;
